Re-write to move most of `rclrs_common` to `rclrs` Distro A, OPSEC #4584. You may have additional rights; please see https://rosmilitary.org/faq/?category=ros-2-license
