+++ rclrs/Cargo.toml
@@ -15,10 +15,11 @@ cty = "0.2"
 core-error = "0.0.0"
 parking_lot = {version = "0.11.2", optional = true}
 spin = "0.9.2"
+downcast = "0.10.0"
 
-[dependencies.rclrs_common]
+[dependencies.rclrs_msg_utilities]
 version = "*"
-path = "../rclrs_common"
+path = "../rclrs_msg_utilities"
 
 [build-dependencies]
 bindgen = "0.59.1"
