+++ rclrs/package.xml
@@ -10,10 +10,9 @@
   <buildtool_depend>ament_cmake</buildtool_depend>
   <buildtool_depend>ament_cmake_export_crates</buildtool_depend>
 
-  <build_depend>builtin_interfaces</build_depend>
+  <build_depend>rclrs_msg_utilities</build_depend>
   <build_depend>rcl</build_depend>
   <build_depend>rcl_interfaces</build_depend>
-  <build_depend>rclrs_common</build_depend>
 
   <export>
     <build_type>ament_cmake</build_type>
