--- rclrs_common/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "rclrs_common"
-version = "0.1.0"
-authors = ["Esteve Fernandez <esteve@apache.org>"]
-edition = "2018"
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-libc = "0.2"
-downcast = {version = "0.10", default-features = false}
-core-error = "0.0.0"
