--- rclrs_common/package.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>rclrs_common</name>
-  <version>0.0.3</version>
-  <description>Common files for the Rust client.</description>
-  <author email="esteve@apache.org">Esteve Fernandez</author>
-  <maintainer email="esteve@apache.org">Esteve Fernandez</maintainer>
-  <license>Apache License 2.0</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  <buildtool_depend>ament_cmake_export_crates</buildtool_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
