+++ rclrs_examples/package.xml
@@ -13,10 +13,12 @@
 
   <build_depend>example_interfaces</build_depend>
   <build_depend>rclrs</build_depend>
+  <build_depend>rclrs_msg_utilities</build_depend>
   <build_depend>std_msgs</build_depend>
 
   <exec_depend>rosidl_generator_rs</exec_depend>
   <exec_depend>rclrs</exec_depend>
+  <exec_depend>rclrs_msg_utilities</exec_depend>
   <exec_depend>std_msgs</exec_depend>
 
   <export>
