+++ rclrs_msg_utilities/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.5)
-project(rclrs NONE)
+project(rclrs_msg_utilities NONE)
 
 find_package(ament_cmake REQUIRED)
 find_package(ament_cmake_export_crates REQUIRED)
@@ -14,7 +14,7 @@ add_custom_target(
 )
 
 install(FILES
-    ${CMAKE_SOURCE_DIR}/target/$<$<NOT:$<CONFIG:Debug>>:release>$<$<CONFIG:Debug>:debug>/librclrs.rlib
+    ${CMAKE_SOURCE_DIR}/target/$<$<NOT:$<CONFIG:Debug>>:release>$<$<CONFIG:Debug>:debug>/librclrs_msg_utilities.rlib
     PERMISSIONS OWNER_READ OWNER_WRITE OWNER_EXECUTE GROUP_READ GROUP_EXECUTE WORLD_READ WORLD_EXECUTE
     DESTINATION lib/${PROJECT_NAME}
 )
