+++ rclrs_msg_utilities/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "rclrs_msg_utilities"
+version = "0.1.0"
+authors = ["Jacob Hassold <jhassold@dcscorp.com"]
+edition = "2018"
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+downcast-rs = "1.2.0"
+libc = "0.2.107"
