+++ rclrs_msg_utilities/package.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0"?>
+<package format="2">
+  <name>rclrs_msg_utilities</name>
+  <version>0.0.1</version>
+  <description>Message generation code shared by Rust projects in ROS 2</description>
+  <author email="jhassold@dcscorp.com">Jacob Hassold</author>
+  <maintainer email="jhassold@dcscorp.com">Jacob Hassold</maintainer>
+  <license>Apache License 2.0</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>ament_cmake_export_crates</buildtool_depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
