+++ rosidl_generator_rs/resource/msg.rs.em
@@ -1,6 +1,6 @@
 use libc::c_char;
 use libc::uintptr_t;
-use rclrs_common;
+use rclrs_msg_utilities;
 use std::ffi::CString;
 use std::ffi::CStr;
 
@@ -92,7 +92,7 @@ impl @(type_name) {
   }
 }
 
-impl rclrs_common::traits::Message for @(type_name) {
+impl rclrs_msg_utilities::traits::Message for @(type_name) {
   fn get_native_message(&self) -> uintptr_t {
     return self.get_native_message();
   }
@@ -106,7 +106,7 @@ impl rclrs_common::traits::Message for @(type_name) {
   }
 }
 
-impl rclrs_common::traits::MessageDefinition<@(type_name)> for @(type_name) {
+impl rclrs_msg_utilities::traits::MessageDefinition<@(type_name)> for @(type_name) {
   fn get_type_support() -> uintptr_t {
     return unsafe { @(package_name)_@(subfolder)_@(convert_camel_case_to_lower_case_underscore(type_name))_get_type_support() };
   }
